Move the compactionjobstatustore
