+++ java/clients/src/main/java/sleeper/clients/status/report/job/query/DetailedJobsQuery.java
@@ -15,7 +15,7 @@
  */
 package sleeper.clients.status.report.job.query;
 
-import sleeper.compaction.core.job.CompactionJobStatusStore;
+import sleeper.core.tracker.compaction.job.CompactionJobStatusStore;
 import sleeper.core.tracker.compaction.job.query.CompactionJobStatus;
 import sleeper.ingest.core.job.status.IngestJobStatus;
 import sleeper.ingest.core.job.status.IngestJobStatusStore;
