+++ java/clients/src/main/java/sleeper/clients/status/report/job/query/UnfinishedJobsQuery.java
@@ -15,8 +15,8 @@
  */
 package sleeper.clients.status.report.job.query;
 
-import sleeper.compaction.core.job.CompactionJobStatusStore;
 import sleeper.core.table.TableStatus;
+import sleeper.core.tracker.compaction.job.CompactionJobStatusStore;
 import sleeper.core.tracker.compaction.job.query.CompactionJobStatus;
 import sleeper.ingest.core.job.status.IngestJobStatus;
 import sleeper.ingest.core.job.status.IngestJobStatusStore;
