+++ java/clients/src/test/java/sleeper/clients/admin/testutils/AdminClientStatusStoreHolder.java
@@ -16,10 +16,10 @@
 package sleeper.clients.admin.testutils;
 
 import sleeper.clients.admin.AdminClientStatusStoreFactory;
-import sleeper.compaction.core.job.CompactionJobStatusStore;
 import sleeper.compaction.core.task.CompactionTaskStatusStore;
 import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.properties.table.TablePropertiesProvider;
+import sleeper.core.tracker.compaction.job.CompactionJobStatusStore;
 import sleeper.ingest.batcher.core.IngestBatcherStore;
 import sleeper.ingest.core.job.status.IngestJobStatusStore;
 import sleeper.ingest.core.task.IngestTaskStatusStore;
