+++ java/core/src/main/java/sleeper/core/tracker/compaction/job/CompactionJobStatusStore.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.core.job;
+package sleeper.core.tracker.compaction.job;
 
 import sleeper.core.tracker.compaction.job.query.CompactionJobStatus;
 import sleeper.core.tracker.compaction.job.update.CompactionJobCommittedEvent;
